Move data package from app module to data module. Refactor ResourceProvider
