+++ app/src/main/java/io/github/msh91/arch/util/providers/BaseResourceProvider.kt
@@ -1,13 +1,10 @@
 package io.github.msh91.arch.util.providers
 
-import android.content.ContentResolver
 import android.graphics.Bitmap
 import android.graphics.drawable.Drawable
-import android.net.Uri
 import androidx.annotation.ColorRes
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
-import java.io.InputStream
 
 /**
  * Resolves application's resources.
@@ -46,32 +43,6 @@ interface BaseResourceProvider {
      * @return Drawable representation of the {@param id}
      */
     fun getDrawable(@DrawableRes resId: Int): Drawable?
-
-    /**
-     * Resolves asset name to to Inputstream
-     *
-     * @param assetName to  fetched from the asset
-     * @return InputStream
-     */
-    fun getAsset(fileName: String): InputStream
-
-    /**
-     * Resolves uri to Inputstream
-     *
-     * @param uri
-     * @return InputStream
-     */
-    fun getContentInputStream(uri: Uri): InputStream
-
-    /**
-     * Get MimType of given [Uri]
-     */
-    fun getMimType(uri: Uri): String
-
-    /**
-     * get an instance of [ContentResolver]
-     */
-    fun getContentResolver(): ContentResolver
 
     /**
      * get an instance of [Bitmap] from given vector resource
