+++ data/build.gradle
@@ -1,0 +1,65 @@
+apply plugin: 'com.android.library'
+apply plugin: 'kotlin-android-extensions'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-kapt'
+
+android {
+    compileSdkVersion 28
+    buildToolsVersion "29.0.1"
+
+
+    defaultConfig {
+        minSdkVersion 16
+        targetSdkVersion 28
+        versionCode 1
+        versionName "1.0"
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+
+    }
+    compileOptions {
+        targetCompatibility 1.8
+        sourceCompatibility 1.8
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        }
+    }
+
+}
+
+kapt {
+    generateStubs true
+    correctErrorTypes true
+}
+
+dependencies {
+    def appDependencies = rootProject.ext.dependencies
+    def testDependencies = rootProject.ext.testDependencies
+    
+    implementation project(':domain')
+
+    implementation appDependencies.kotlin
+    implementation appDependencies.ArchLifeCycle
+
+    implementation appDependencies.dagger
+    kapt appDependencies.daggerCompiler
+
+    implementation appDependencies.Room
+    implementation appDependencies.RxRoom
+    kapt appDependencies.RoomCompiler
+
+    implementation appDependencies.rxJava
+    implementation appDependencies.rxAndroid
+
+    implementation appDependencies.gson
+    implementation appDependencies.retrofit
+    implementation appDependencies.retrofitGson
+    implementation appDependencies.retrofitRx
+
+    testImplementation testDependencies.junit
+    testImplementation testDependencies.mockitoKotlin
+}
