+++ data/src/main/java/io/github/msh91/arch/data/repository/movie/MovieRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package io.github.msh91.arch.data.repository.movie
 
+import io.github.msh91.arch.data.di.qualifier.network.Mock
 import io.github.msh91.arch.data.source.cloud.BaseCloudRepository
 import io.github.msh91.arch.data.source.db.dao.MovieDao
-import io.github.msh91.arch.di.qualifier.network.Mock
 import io.github.msh91.arch.domain.model.movie.Movie
 import io.github.msh91.arch.domain.repository.MovieRepository
 import io.reactivex.Flowable
