+++ data/src/main/java/io/github/msh91/arch/data/source/cloud/BaseCloudRepository.kt
